backends/core/requirements.txt
Flask==0.12.2
Flask-Cors==3.0.2
Flask-Migrate==2.0.3
Flask-SQLAlchemy==2.2
MySQL-python==1.2.5
sqlalchemy_mixins==0.2.2
google-cloud-bigquery==0.27
simpleeval==0.9.5
Flask-RESTful==0.3.5
google-api-python-client==1.6.5
google-cloud-logging==1.6.0
GoogleAppEngineCloudStorageClient==1.9.22.1
pyyaml==3.12
